refactor blits, less copy+paste with slight modifications
this might still need to be tweaked still as benchmarks are showing
some performance loses (though, also, some minor performance gains?)
so i guess the lambda functions may not always be inlined, or if they
are, the optimizer isn't able to do the same level of optimizations as
it was before? unsure as of yet, so need to investigate further.

however, none of the performance regressions were huge, so i think this
is acceptable to me for now.

libretrogd was never intended to push performance boundaries anyway :-)