convert sdl2 events to our own equivalents
this is to not force _some_ sdl2 type usage onto applications, thus
forcing them to also include sdl2 as a dependency themselves.

that being said, this is still pretty gross feeling to me and i wish
there was a better solution.

i've not added all possible sdl2 event equivalents. only the ones that
i am currently interested in and use. i'll add more as is needed. likely
the joystick / controller events will be added in the very near future.