remove Copy trait from these event enums to be consistent with the rest
since the TextInput event cannot implement Copy, necessitating the
removal of the Copy trait from KeyboardEvent, it feels wrong to me to
have the other event enums implement Copy, because it is inconsistent.

so while this is less convenient potentially, it feels consistent.

probably not a big deal in the future one way or the other.