specify nightly toolchain by date
i did not expect this to be problematic so soon, based on what i'd
read in various places about the stability of nightly (especially so
for the most common platforms like x86_64, etc)

as i write this commit message, all nightly versions past this date
fail to build ggdt, failing with an "illegal instruction" error.

sigh. this makes me question whether trying to use simd right now is
actually worth it ...